requirements.txt
google-adk>=1.19.0
supabase>=2.9.0
